backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
pdfminer.six
duckduckgo-search
